Tests/LibWeb/Text/expected/Worker/Worker-postMessage-transfer.txt

Message received from worker: Hello, world
